+++ .ariesrc.js
@@ -3,7 +3,7 @@ module.exports = {
   url: './test/swagger.json',
   autoMock: true,
   resTemplate: '{"code":0,"msg":"success","data":$data}',
-  defaultRequired: true,
+  autoRequired: true,
   formatProp: (prop) => {
     // Add prop format
     return prop